api/requirements.txt
flask-restx
lxml
GDAL
